Create solution to day 1 problem 1
Solve the day 1 problem 1 by grouping into a set of pairs and
enumerating them. For example, in the test case, [199, 200, 208...], the pairs
would be (199, 200), (200, 208), etc. This makes it trivial to run over
them with .filter to find where the number has increased, and count the
results.

Most of this change is tooling to ensure the dev environment is set up
correctly. I've added FsUnit for idiomatic tests, and Paket for
idiomatic package management.